Release v0.6.2 (#53)
* Updated to new changelog system (#32)

* fix: added new changelog system

* fix: updated license file

* fix: added bug report and feature request templates (#33)

* feature: improved naming scheme in output files (#35)

Names are now more intuitive and human-readable. They also include units
whenever possible.

* fix: added outdated notice to README file (#36)

* Switched TODO list to markdown format (#37)

* fix: switched to-do list to markdown format

* fix: removed old release notes file

* fix: updated changelog

Failed to add latest feature to file.

* Added flux-converted photometry to the output files (#38)

* feat: photometry flux are now included in the output files

Also, reimplemented the output naming scheme, because I'm an idiot

* fix: update TODO list for feature

* fix: tweaked README layout

* fix: updated changelog

* Release-v0.6.0 (#39)

* fix: prepared changelog for release

* fix: updated TODO list

* PushChangelog (#42)

* Release: v0.6.0-beta.5 (#40)

* Updated to new changelog system (#32)

* fix: added new changelog system

* fix: updated license file

* fix: added bug report and feature request templates (#33)

* feature: improved naming scheme in output files (#35)

Names are now more intuitive and human-readable. They also include units
whenever possible.

* fix: added outdated notice to README file (#36)

* Switched TODO list to markdown format (#37)

* fix: switched to-do list to markdown format

* fix: removed old release notes file

* fix: updated changelog

Failed to add latest feature to file.

* Added flux-converted photometry to the output files (#38)

* feat: photometry flux are now included in the output files

Also, reimplemented the output naming scheme, because I'm an idiot

* fix: update TODO list for feature

* fix: tweaked README layout

* fix: updated changelog

* Release-v0.6.0 (#39)

* fix: prepared changelog for release

* fix: updated TODO list

* docs(changelog): Update CHANGELOG

Co-authored-by: ImaanBontle <ImaanBontle@users.noreply.github.com>

* fix: added supported for different OS directory schemes (#43)

Custom case-switches are added for each type of OS. User will be
prompted to create a custom case-switch if their system is not
recognised.

* fix: added readme file for output files (#44)

* fix: fixed incorrect flux conversion in the output (#45)

fix: updated README to reflect changes

* docs(changelog): Update CHANGELOG

* fix: fixed bug with source RA and Dec in summary files (#48)

Source RA and Dec were not correctly propagating in the summary files,
resulting in the same values for all galaxies. This has now been fixed.

* Updated workflows (#50)

* fix: tweaked release-drafter syntax

* fix: removed copy-changelog.yml

Changelog should only be updated to main when ready for release, in
agreement with gitflow best practice,

* fix: release-drafter should only watch main now

This is to ensure the changelog is based on the main branch and that the
ref tag is set to main as well

* fix: update-changelog will wait 2 minutes for checks to complete

This is to avoid accidentally deleting the branch before checks have
completed

* fix: added sorting to output files (#51)

* fix: renamed FILES.md to README.md

This is to make it easier to notice when appended with output files

* fix: updated README for v0.6.2

Co-authored-by: ImaanBontle <ImaanBontle@users.noreply.github.com>